requirements.txt
discord.py
deep-translator
Flask
